+++ TinkoffASDKUI/TinkoffASDKUITests/TestCases/BankResolverTests.swift
@@ -10,9 +10,11 @@ import Foundation
 @testable import TinkoffASDKUI
 import XCTest
 
-final class BankResolverTests: XCTestCase {
+final class BankResolverTests: BaseTestCase {
 
     func test_resolve_should_return_noValue_error() throws {
+        allureId(2416099, "BankResolver возвращает ошибку noValue для пустой строки")
+
         let sut = getSut()
         let expectedResult = BankResult.incorrectInput(error: .noValue)
         let result = sut.resolve(cardNumber: "")
@@ -20,9 +22,20 @@ final class BankResolverTests: XCTestCase {
     }
 
     func test_resolve_should_return_notEnoughData_error() throws {
+        allureId(2416094, "BankResolver возвращает ошибку notEnoughData для BIN < 6 символов")
+
         let sut = getSut()
         let expectedResult = BankResult.incorrectInput(error: .notEnoughData)
         let result = sut.resolve(cardNumber: "1234")
+        XCTAssertEqual(result, expectedResult)
+    }
+
+    func test_resolve_should_return_noValue_error_when_value_is_nil() {
+        allureId(2416116, "BankResolver возвращает ошибку noValue для nil")
+
+        let sut = getSut()
+        let expectedResult = BankResult.incorrectInput(error: .noValue)
+        let result = sut.resolve(cardNumber: nil)
         XCTAssertEqual(result, expectedResult)
     }
 
@@ -55,6 +68,8 @@ private extension BankResolverTests {
     // MARK: - Sub test functions
 
     func resolve_should_return_parsed_alpha() throws {
+        allureId(2415829, "BankResolver определяет BIN для Альфа")
+
         // given
         let sut = getSut()
         let expectedResult = BankResult.parsed(bank: .alpha)
@@ -70,6 +85,8 @@ private extension BankResolverTests {
     }
 
     func resolve_should_return_parsed_sber() throws {
+        allureId(2415806, "BankResolver определяет BIN для Сбер")
+
         // given
         let sut = getSut()
         let expectedResult = BankResult.parsed(bank: .sber)
@@ -85,6 +102,8 @@ private extension BankResolverTests {
     }
 
     func resolve_should_return_parsed_vtb() throws {
+        allureId(2415839, "BankResolver определяет BIN для ВТБ")
+
         // given
         let sut = getSut()
         let expectedResult = BankResult.parsed(bank: .vtb)
@@ -100,6 +119,8 @@ private extension BankResolverTests {
     }
 
     func resolve_should_return_parsed_tinkoff() throws {
+        allureId(2415809, "BankResolver определяет BIN для Тинькофф")
+
         // given
         let sut = getSut()
         let expectedResult = BankResult.parsed(bank: .tinkoff)
@@ -115,6 +136,8 @@ private extension BankResolverTests {
     }
 
     func resolve_should_return_parsed_gazprom() throws {
+        allureId(2415807, "BankResolver определяет BIN для Газпром")
+
         // given
         let sut = getSut()
         let expectedResult = BankResult.parsed(bank: .gazprom)
@@ -130,6 +153,8 @@ private extension BankResolverTests {
     }
 
     func resolve_should_return_parsed_raifaissen() throws {
+        allureId(2415815, "BankResolver определяет BIN для Райфайзен")
+
         // given
         let sut = getSut()
         let expectedResult = BankResult.parsed(bank: .raiffaisen)
@@ -145,6 +170,8 @@ private extension BankResolverTests {
     }
 
     func resolve_should_return_parsed_other() throws {
+        allureId(2415840, "BankResolver определяет BIN для другого банка")
+
         // given
         let sut = getSut()
         let expectedResult = BankResult.parsed(bank: .other)